splitting redux folder into reducers and actions folders
