+++ src/index.js
@@ -7,7 +7,7 @@ import registerServiceWorker from './registerServiceWorker';
 
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
-import reducers from './redux/reducers';
+import reducers from './reducers/reducers';
 
 let store = createStore(reducers, window.__REDUX_DEVTOOLS_EXTENSION__ &&
   window.__REDUX_DEVTOOLS_EXTENSION__());
