+++ src/reducers/reducers.js
@@ -1,11 +1,11 @@
 import { combineReducers } from 'redux';
-import * as fromAuth from './authReducer';
+import authReducer, * as fromAuth from './authReducer';
 import {
   ON_DETECTED,
   ON_SEARCH,
   LOG_IN,
   ADD_BOOK
-} from './actionTypes'
+} from '../actions/actionTypes'
 
 const defaultState = {
   results: {},
@@ -58,8 +58,13 @@ const userReducer = (state = defaultUserState, action) => {
 
 // Combining all reducers.
 export default combineReducers({
+  auth: authReducer,
+
   onDetectedReducer,
   userReducer
 });
 
-export const isAuthenticated = state => fromAuth.isAuthenticated(state.auth);
+export const isAuthenticated = state => {
+  console.log(state);
+  return fromAuth.isAuthenticated(state.auth);
+}
